LPS-86616 Move the logic to unit test
