+++ portal-impl/test/unit/com/liferay/portal/util/PortalImplUnitTest.java
@@ -254,6 +254,14 @@ public class PortalImplUnitTest {
 	}
 
 	@Test
+	public void testGetHostWithIPv6Address() {
+		_assertGetHost("[::1]", "::1");
+		_assertGetHost("[::1]:80", "::1");
+		_assertGetHost("[0:0:0:0:0:0:0:1]", "0:0:0:0:0:0:0:1");
+		_assertGetHost("[0:0:0:0:0:0:0:1]:80", "0:0:0:0:0:0:0:1");
+	}
+
+	@Test
 	public void testGetOriginalServletRequest() {
 		HttpServletRequest httpServletRequest = new MockHttpServletRequest();
 
@@ -685,6 +693,15 @@ public class PortalImplUnitTest {
 		throws Exception {
 
 		ReflectionTestUtil.setFieldValue(PropsValues.class, fieldName, value);
+	}
+
+	private void _assertGetHost(String httpHostHeader, String host) {
+		MockHttpServletRequest mockHttpServletRequest =
+			new MockHttpServletRequest();
+
+		mockHttpServletRequest.addHeader("Host", httpHostHeader);
+
+		Assert.assertEquals(host, _portalImpl.getHost(mockHttpServletRequest));
 	}
 
 	private final PortalImpl _portalImpl = new PortalImpl();
